refactor: view, model 패키지
