--- src/main/java/racing/NumberProvider.java
@@ -1,5 +1,0 @@
-package racing;
-
-interface NumberProvider {
-    int getNumber();
-}
