+++ src/main/java/racing/domain/Car.java
@@ -1,5 +1,4 @@
-package racing;
-
+package racing.domain;
 
 
 public class Car {
@@ -7,8 +6,6 @@ public class Car {
     private Position curPosition;
     private final MovableStrategy movableStrategy;
     private CarName name;
-
-    private static final String CAR_STATE_BAR = "-";
 
     public Car(MovableStrategy movableStrategy, String name) {
         this.curPosition = new Position();
@@ -30,13 +27,7 @@ public class Car {
         return curPosition.getPosition();
     }
 
-    public String getCurMoveStateExp(){
-        StringBuilder sb = new StringBuilder();
-        for (int i = 0; i < getCurPosition(); i++) {
-            sb.append(CAR_STATE_BAR);
-        }
-        return sb.toString();
-    }
+
 
     public String getName(){
         return name.getName();
