+++ src/main/java/racing/domain/CarName.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain;
 
 public class CarName {
     private final String name;
