+++ src/main/java/racing/domain/MovableStrategy.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain;
 
 public interface MovableStrategy {
     boolean canMove();
