+++ src/main/java/racing/domain/NumberProvider.java
@@ -1,0 +1,5 @@
+package racing.domain;
+
+public interface NumberProvider {
+    int getNumber();
+}
