+++ src/main/java/racing/domain/Position.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain;
 
 public class Position {
