+++ src/main/java/racing/domain/RandomNumMovableStrategy.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain;
 
 public class RandomNumMovableStrategy implements MovableStrategy {
