+++ src/test/java/racing/CarsTest.java
@@ -3,6 +3,7 @@ package racing;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racing.domain.*;
 
 import java.util.ArrayList;
 import java.util.List;
