+++ src/test/java/racing/NameTest.java
@@ -2,6 +2,9 @@ package racing;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racing.domain.Car;
+import racing.domain.MovableStrategy;
+import racing.domain.RandomNumMovableStrategy;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
